The information in argument rdayvrai of calfis was redundant with the information in argument time. Furthermore, in the physics part of gcm, we need separately the day number (an integer) and the time of day. So, replaced real argument rdayvrai of calfis containing elapsed time by integer argument dayvrai containing day number. Corresponding change in leapfrog. In procedure physiq, replaced real argument rdayvrai by integer argument dayvrai. In procedures readsulfate and readsulfate_preind, replaced real argument r_day by arguments dayvrai and time.
In procedure alboc, replaced real argument rjour by integer argument
jour. alboc was always called by interfsurf_hq with actual argument
real(jour), and the meaning of the dummy argument in alboc seems to be
that it should be an integer.

In procedure leapfrog, local variable time could not be > 1. Removed
test.

In physiq, replaced nint(rdayvrai) by dayvrai. This changes the
results since julien now changes at 0 h instead of 12 h. This follows
LMDZ, where the argument of ozonecm is days_elapsed+1.

+++ phylmd/Interface_surf/interfsurf_hq.f
@@ -276,7 +276,7 @@ contains
 
        ! calcul albedo
        if (minval(rmu0) == maxval(rmu0) .and. minval(rmu0) == -999.999) then
-          CALL alboc(FLOAT(jour), rlat, alb_eau)
+          CALL alboc(jour, rlat, alb_eau)
        else ! cycle diurne
           CALL alboc_cd(rmu0, alb_eau)
        endif
